pzem_server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyserial==3.5
python-dotenv==1.0.0
psycopg2-binary